requirements.txt
python-dotenv==1.0.0
tweepy==4.14.0
PyYAML==6.0.1
pyserial==3.5
keyboard==0.13.5
colorama==0.4.6
